Organize flight import wizard into separate files for separate wizard pages
A long time coming
