--- src/main/python/importers/__init__.py
@@ -1,3 +1,0 @@
-
-from .flight import FlightImportWizard
-
