+++ src/main/python/importwizards/__init__.py
@@ -1,0 +1,2 @@
+
+from .flightimportwizard import FlightImportWizard
