+++ src/main/python/importwizards/pages/__init__.py
@@ -1,0 +1,9 @@
+
+from .ids import PageIds
+
+from .intro import IntroPage
+from .meta import MetadataPage
+from .parameters import ParametersPage
+from .setlibrary import SetLibraryPage
+from .review import ReviewPage
+from .conclusion import ConclusionPage
