+++ src/main/python/importwizards/pages/ids.py
@@ -1,0 +1,8 @@
+
+class PageIds:
+    Page_Intro = 1
+    Page_Parameters = 2
+    Page_Review = 3
+    Page_Metadata = 4
+    Page_SetLibrary = 5
+    Page_Conclusion = 6
